Rename tracking_field_bitbucket to tracking_field.
