+++ scilpy/tracking/utils.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+
+
+class TrackingDirection(list):
+
+    """
+    Tracking direction use as 3D cartesian direction (list(x,y,z))
+    and has an index to work with discrete sphere.
+    """
+
+    def __init__(self, cartesian, index=None):
+        super(TrackingDirection, self).__init__(cartesian)
+        self.index = index
